@@ -1,16 +1,26 @@
-apply plugin: 'com.android.application'
+
+plugins {
+    id 'com.android.application'
+    id 'kotlin-android'
+}
+apply plugin: 'com.jakewharton.hugo'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.2'
+    compileSdkVersion 30
 
     defaultConfig {
         applicationId "com.tzutalin.dlibtest"
         minSdkVersion 21
-        targetSdkVersion 25
+        targetSdkVersion 30
         versionCode 1
         versionName "${rootProject.ext.releaseVersionName}"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+        javaCompileOptions {
+            annotationProcessorOptions {
+                arguments = ["androidManifestFile": "$projectDir/src/main/AndroidManifest.xml".toString()]
+            }
+        }
     }
     buildTypes {
         release {
@@ -18,38 +28,49 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
+    }
+
+
+
+
+
 }
 
 
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath 'com.neenbedankt.gradle.plugins:android-apt:1.4'
-    }
-}
 
 repositories {
     mavenCentral()
     mavenLocal()
 }
 
-apply plugin: 'com.neenbedankt.android-apt'
-def AAVersion = '4.0.0'
+//apply plugin: 'com.neenbedankt.android-apt'
+def AAVersion = '4.6.0'
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile "com.android.support:appcompat-v7:${rootProject.ext.androidSupportSdkVersion}"
-    compile 'com.android.support:design:25.2.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'com.google.android.material:material:1.3.0'
     compile 'com.github.dexafree:materiallist:3.0.1'
     compile 'com.jakewharton.timber:timber:4.5.1'
     compile project(':dlib')
-    apt "org.androidannotations:androidannotations:$AAVersion"
-    compile "org.androidannotations:androidannotations-api:$AAVersion"
+
+    annotationProcessor "org.androidannotations:androidannotations:$AAVersion"
+    implementation "org.androidannotations:androidannotations-api:$AAVersion"
+
+
+    testImplementation group: 'org.assertj', name: 'assertj-core', version: '3.19.0'
 
     // Add AndroidJUnit
-    androidTestCompile "com.android.support:support-annotations:${rootProject.ext.androidSupportSdkVersion}"
+    testImplementation 'junit:junit:4.12'
+
+//    androidTestCompile "com.android.support:support-annotations:29.0.1"
     androidTestCompile 'com.android.support.test:runner:0.5'
     androidTestCompile 'com.android.support.test:rules:0.5'
     // Optional -- Hamcrest library
