@@ -5,12 +5,11 @@ version = "${rootProject.ext.releaseVersionName}"
 group = "com.tzutalin.dlib-android-app"
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.2'
+    compileSdkVersion 30
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 25
+        targetSdkVersion 30
         versionCode 1
         versionName "${rootProject.ext.releaseVersionName}"
 
@@ -47,10 +46,10 @@ configurations {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile "com.android.support:appcompat-v7:${rootProject.ext.androidSupportSdkVersion}"
-    javadocDeps "com.android.support:appcompat-v7:${rootProject.ext.androidSupportSdkVersion}"
-    compile "com.android.support:support-annotations:${rootProject.ext.androidSupportSdkVersion}"
-    javadocDeps "com.android.support:support-annotations:${rootProject.ext.androidSupportSdkVersion}"
+
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    javadocDeps "androidx.appcompat:appcompat:1.2.0"
+
 }
 
 buildscript {
