@@ -17,8 +17,9 @@
 package com.tzutalin.dlibtest;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.RawRes;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.RawRes;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
