@@ -19,10 +19,11 @@ package com.tzutalin.dlibtest;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Paint;
-import android.support.annotation.NonNull;
 import android.util.AttributeSet;
 import android.util.TypedValue;
 import android.view.View;
+
+import androidx.annotation.NonNull;
 
 public class TrasparentTitleView extends View {
     private static final float TEXT_SIZE_DIP = 24;
